# docs/swagger.yaml
basePath: /api
definitions:
...
  myobj_src_core_domain_request.UserLoginRequest:
    properties:
      challenge:
        type: string
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  myobj_src_core_domain_request.UserRegisterRequest:
    properties:
      challenge:
        description: 挑战ID
        type: string
      email:
        type: string
      nickname:
        type: string
      password:
        type: string
      phone:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  myobj_src_core_domain_request.UserUpdatePasswordRequest:
    properties:
      challenge:
        description: 挑战ID
        type: string
      id:
        type: string
      new_passwd:
        type: string
      old_passwd:
        type: string
    type: object
  myobj_src_core_domain_request.UserUpdateRequest:
    properties:
      email:
        description: 用户邮箱
        type: string
      id:
        type: string
      nickname:
        description: 用户昵称
        type: string
      phone:
        description: 用户手机号
        type: string
      username:
        description: 用户名
        type: string
    type: object
  myobj_src_pkg_models.JsonResponse:
    properties:
      code:
        description: 状态码
        type: integer
      data:
        description: 数据
      message:
        description: 状态
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@myobj.com
    name: API Support
    url: https://github.com/yourusername/myobj/issues
  description: |-
    MyObj 是一个功能强大的文件存储系统，支持文件上传、下载、分享、回收站等功能
    支持大文件分片上传、秒传、文件加密等高级特性
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://github.com/yourusername/myobj
  title: MyObj 文件存储系统 API
  version: "1.0"
paths:
...
  /user/challenge:
    get:
      description: 获取RSA公钥和挑战值，用于密码加密传输
      produces:
      - application/json
      responses:
        "200":
          description: 挑战值信息
          schema:
            allOf:
            - $ref: '#/definitions/myobj_src_pkg_models.JsonResponse'
            - properties:
                data:
                  type: object
              type: object
        "400":
          description: 获取失败
          schema:
            $ref: '#/definitions/myobj_src_pkg_models.JsonResponse'
      summary: 获取登录挑战值
      tags:
      - 用户管理
  /user/login:
    post:
      consumes:
      - application/json
      description: 用户通过用户名和密码登录系统，返回JWT Token
      parameters:
      - description: 登录请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/myobj_src_core_domain_request.UserLoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: token
          schema:
            allOf:
            - $ref: '#/definitions/myobj_src_pkg_models.JsonResponse'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: 参数错误或登录失败
          schema:
            $ref: '#/definitions/myobj_src_pkg_models.JsonResponse'
      summary: 用户登录
      tags:
      - 用户管理
  /user/register:
    post:
      consumes:
      - application/json
      description: 注册新用户账号
      parameters:
      - description: 注册请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/myobj_src_core_domain_request.UserRegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 注册成功
          schema:
            $ref: '#/definitions/myobj_src_pkg_models.JsonResponse'
        "400":
          description: 参数错误或注册失败
          schema:
            $ref: '#/definitions/myobj_src_pkg_models.JsonResponse'
      summary: 用户注册
      tags:
      - 用户管理
  /user/setFilePassword:
    post:
      consumes:
      - application/json
      description: 设置用户文件加密密码（首次设置）
      parameters:
      - description: 设置文件密码请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/myobj_src_core_domain_request.UserUpdatePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 设置成功
          schema:
            $ref: '#/definitions/myobj_src_pkg_models.JsonResponse'
        "400":
          description: 参数错误或设置失败
          schema:
            $ref: '#/definitions/myobj_src_pkg_models.JsonResponse'
      security:
      - BearerAuth: []
      summary: 设置文件加密密码
      tags:
      - 用户管理
  /user/sysInfo:
    get:
      description: 获取系统配置和初始化状态
      produces:
      - application/json
      responses:
        "200":
          description: 系统信息
          schema:
            $ref: '#/definitions/myobj_src_pkg_models.JsonResponse'
        "400":
          description: 获取失败
          schema:
            $ref: '#/definitions/myobj_src_pkg_models.JsonResponse'
      summary: 获取系统初始化信息
      tags:
      - 系统管理
  /user/updateFilePassword:
    post:
      consumes:
      - application/json
      description: 修改用户文件加密密码
      parameters:
      - description: 修改文件密码请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/myobj_src_core_domain_request.UserUpdatePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 修改成功
          schema:
            $ref: '#/definitions/myobj_src_pkg_models.JsonResponse'
        "400":
          description: 参数错误或修改失败
          schema:
            $ref: '#/definitions/myobj_src_pkg_models.JsonResponse'
      security:
      - BearerAuth: []
      summary: 修改文件加密密码
      tags:
      - 用户管理
  /user/updatePassword:
    post:
      consumes:
      - application/json
      description: 修改用户登录密码
      parameters:
      - description: 修改密码请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/myobj_src_core_domain_request.UserUpdatePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 修改成功
          schema:
            $ref: '#/definitions/myobj_src_pkg_models.JsonResponse'
        "400":
          description: 参数错误或修改失败
          schema:
            $ref: '#/definitions/myobj_src_pkg_models.JsonResponse'
      security:
      - BearerAuth: []
      summary: 修改登录密码
      tags:
      - 用户管理
  /user/updateUser:
    post:
      consumes:
      - application/json
      description: 更新当前用户的基本信息
      parameters:
      - description: 更新请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/myobj_src_core_domain_request.UserUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            $ref: '#/definitions/myobj_src_pkg_models.JsonResponse'
        "400":
          description: 参数错误或更新失败
          schema:
            $ref: '#/definitions/myobj_src_pkg_models.JsonResponse'
      security:
      - BearerAuth: []
      summary: 更新用户信息
      tags:
      - 用户管理
securityDefinitions:
  ApiKeyAuth:
    description: API Key 身份验证
    in: header
    name: X-API-Key
    type: apiKey
  BearerAuth:
    description: 输入 "Bearer {token}" 进行身份验证
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"